fix rank

+++ src/main/kotlin/com/wafflestudio/snugo/record/model/RankResponse.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.snugo.record.model
 
 data class RankResponse(
-	val myRank: Long,
-	val total: Long
+	val myRank: Int,
+	val total: Int
 )
